Get multiple years of EIA 176 CSV data
