+++ src/pudl/extract/eia176.py
@@ -1,26 +1,34 @@
-"""Extract EIA Form 176 data from CSVs.
+"""Extract EIA Form 176 data from CSVs."""
 
-The EIA Form 176 archive also contains CSVs for EIA Form 191 and EIA Form 757.
-"""
+from dagster import AssetOut, Output, multi_asset
 
-from dagster import asset
+from pudl.extract.csv import CsvExtractor
+from pudl.extract.extractor import raw_df_factory
 
-from pudl.extract.csv import CsvExtractor, get_table_file_map
-
-DATASET = "eia176"
+raw_table_names = "raw_eia176__data"
 
 
-@asset(required_resource_keys={"datastore"})
-def raw_eia176__company(context):
+raw_eia176__all_dfs = raw_df_factory(CsvExtractor, name="eia176")
+
+
+@multi_asset(
+    outs={table_name: AssetOut() for table_name in sorted(raw_table_names)},
+    required_resource_keys={"datastore", "dataset_settings"},
+)
+def raw_eia176__data(raw_eia176__all_dfs):
     """Extract raw EIA company data from CSV sheets into dataframes.
 
-    Args:
-        context: dagster keyword that provides access to resources and config.
-
     Returns:
-        An extracted EIA dataframe with company data.
+        A tuple of extracted EIA dataframes.
     """
-    table_file_map = get_table_file_map(DATASET)
-    with context.resources.datastore.get_zipfile_resource(DATASET) as zf:
-        extractor = CsvExtractor(zf, table_file_map)
-        return extractor.extract_one("company")
+    # create descriptive table_names
+    raw_eia176__all_dfs = {
+        "raw_eia860__" + table_name: df
+        for table_name, df in raw_eia176__all_dfs.items()
+    }
+    raw_eia176__all_dfs = dict(sorted(raw_eia176__all_dfs.items()))
+
+    return (
+        Output(output_name=table_name, value=df)
+        for table_name, df in raw_eia176__all_dfs.items()
+    )
