+++ src/pudl/extract/eia860.py
@@ -10,13 +10,14 @@ from dagster import AssetOut, Output, multi_asset
 import pudl
 import pudl.logging_helpers
 from pudl.extract import excel
+from pudl.extract.extractor import raw_df_factory
 from pudl.helpers import remove_leading_zeros_from_numeric_strings
 from pudl.metadata.classes import DataSource
 
 logger = pudl.logging_helpers.get_logger(__name__)
 
 
-class Extractor(excel.GenericExtractor):
+class Extractor(excel.ExcelExtractor):
     """Extractor for the excel dataset EIA860."""
 
     def __init__(self, *args, **kwargs):
@@ -85,7 +86,7 @@ raw_table_names = (
 )
 
 
-raw_eia860__all_dfs = excel.raw_df_factory(Extractor, name="eia860")
+raw_eia860__all_dfs = raw_df_factory(Extractor, name="eia860")
 
 
 # TODO (bendnorman): Figure out type hint for context keyword and mutli_asset return
