+++ src/pudl/extract/eia923.py
@@ -10,12 +10,13 @@ from dagster import AssetOut, Output, multi_asset
 
 import pudl.logging_helpers
 from pudl.extract import excel
+from pudl.extract.extractor import raw_df_factory
 from pudl.helpers import remove_leading_zeros_from_numeric_strings
 
 logger = pudl.logging_helpers.get_logger(__name__)
 
 
-class Extractor(excel.GenericExtractor):
+class Extractor(excel.ExcelExtractor):
     """Extractor for EIA form 923."""
 
     def __init__(self, *args, **kwargs):
@@ -104,7 +105,7 @@ eia_raw_table_names = (
 )
 
 
-raw_eia923__all_dfs = excel.raw_df_factory(Extractor, name="eia923")
+raw_eia923__all_dfs = raw_df_factory(Extractor, name="eia923")
 
 
 # TODO (bendnorman): Figure out type hint for context keyword and mutli_asset return
