+++ test/unit/extract/csv_test.py
@@ -1,52 +1,46 @@
 """Unit tests for pudl.extract.csv module."""
 from unittest.mock import MagicMock, patch
 
-from pudl.extract.csv import CsvExtractor, get_table_file_map, open_csv_resource
+import pandas as pd
+
+from pudl.extract.csv import CsvExtractor
 
 DATASET = "eia176"
-BASE_FILENAME = "table_file_map.csv"
-TABLE_NAME = "company"
-FILENAME = "all_company_176.csv"
-TABLE_FILE_MAP = {TABLE_NAME: FILENAME}
+PARTITION = 2023
+CSV_FILENAME = f"{DATASET}-{PARTITION}.csv"
+PAGE = "data"
 
 
 def get_csv_extractor():
-    zipfile = MagicMock()
-    return CsvExtractor(zipfile, TABLE_FILE_MAP)
+    mock_ds = MagicMock()
+    return CsvExtractor(mock_ds, DATASET)
 
 
-def test_open_csv_resource():
-    csv_resource = open_csv_resource(DATASET, BASE_FILENAME)
-    assert ["table", "filename"] == csv_resource.fieldnames
-
-
-def test_get_table_file_map():
-    table_file_map = get_table_file_map(DATASET)
-    assert table_file_map == TABLE_FILE_MAP
-
-
-def test_get_table_names():
+def test_csv_filename():
     extractor = get_csv_extractor()
-    table_names = extractor.get_table_names()
-    assert [TABLE_NAME] == table_names
+    assert extractor.csv_filename(PARTITION) == CSV_FILENAME
 
 
 @patch("pudl.extract.csv.pd")
-def test_csv_extractor_read_source(mock_pd):
+def test_load_csv_file(mock_pd):
     extractor = get_csv_extractor()
-    res = extractor.extract_one(TABLE_NAME)
-    mock_zipfile = extractor._zipfile
-    mock_zipfile.open.assert_called_once_with(FILENAME)
-    f = mock_zipfile.open.return_value.__enter__.return_value
-    mock_pd.read_csv.assert_called_once_with(f)
-    df = mock_pd.read_csv()
-    assert df == res
+
+    assert mock_pd.read_csv.return_value == extractor.load_csv_file(year=PARTITION)
+    extractor.ds.get_zipfile_resource.assert_called_once_with(DATASET)
+    zipfile = extractor.ds.get_zipfile_resource.return_value.__enter__.return_value
+    zipfile.open.assert_called_once_with(CSV_FILENAME)
+    file = zipfile.open.return_value.__enter__.return_value
+    mock_pd.read_csv.assert_called_once_with(file)
 
 
-def test_csv_extractor_extract():
+def test_extract():
     extractor = get_csv_extractor()
-    df = MagicMock()
-    with patch.object(CsvExtractor, "extract_one", return_value=df) as mock_read_source:
-        raw_dfs = extractor.extract_all()
-    mock_read_source.assert_called_once_with(TABLE_NAME)
-    assert {TABLE_NAME: df} == raw_dfs
+    dummy_field = "company"
+    dummy_value = "Total of All Companies"
+    df = pd.DataFrame([dummy_value])
+    df.columns = [dummy_field]
+    with patch.object(CsvExtractor, "load_csv_file", return_value=df):
+        res = extractor.extract(year=PARTITION)
+    assert len(res) == 1
+    print(res)
+    assert dummy_value == res[PAGE][dummy_field][0]
