+++ test/unit/extract/excel_test.py
@@ -4,11 +4,8 @@ from unittest import mock as mock
 from unittest.mock import patch
 
 import pandas as pd
-import pytest
-from dagster import build_op_context
 
-from pudl.extract import excel as excel
-from pudl.settings import DatasetsSettings
+from pudl.extract import excel
 
 
 class TestMetadata(unittest.TestCase):
@@ -35,7 +32,7 @@ class TestMetadata(unittest.TestCase):
         self.assertEqual(1, self._metadata.get_sheet_name("boxes", year=2011))
 
 
-class FakeExtractor(excel.GenericExtractor):
+class FakeExtractor(excel.ExcelExtractor):
     """Test friendly fake extractor returns strings instead of files."""
 
     def __init__(self, *args, **kwargs):
@@ -79,8 +76,8 @@ def _fake_data_frames(page_name, **kwargs):
     return fake_data[page_name]
 
 
-class TestGenericExtractor(unittest.TestCase):
-    """Test operation of the excel.GenericExtractor class."""
+class TestExtractor(unittest.TestCase):
+    """Test operation of the excel.Extractor class."""
 
     @staticmethod
     @patch("pudl.extract.excel.pd.read_excel")
@@ -118,39 +115,3 @@ class TestGenericExtractor(unittest.TestCase):
 
     # TODO(rousik@gmail.com): need to figure out how to test process_$x methods.
     # TODO(rousik@gmail.com): we should test that empty columns are properly added.
-
-
-@pytest.mark.parametrize(
-    "dataset, expected_years",
-    (
-        ("eia860", set(range(2001, 2022))),
-        ("eia861", set(range(2001, 2022))),
-        ("eia923", set(range(2001, 2022))),
-    ),
-)
-def test_years_from_settings(dataset, expected_years):
-    years_from_settings = excel.years_from_settings_factory(dataset)
-
-    with build_op_context(
-        resources={"dataset_settings": DatasetsSettings()}
-    ) as context:
-        # Assert actual years are a superset of expected. Instead of doing
-        # an equality check, this avoids having to update expected years
-        # every time a new year is added to the datasets
-        assert {
-            output.value for output in years_from_settings(context)
-        } >= expected_years
-
-
-def test_concat_pages():
-    pages = ["page1", "page2", "page3"]
-    dfs_1 = {page: pd.DataFrame({"df": [1], "page": [page]}) for page in pages}
-    dfs_2 = {page: pd.DataFrame({"df": [2], "page": [page]}) for page in pages}
-
-    merged_dfs = excel.concat_pages([dfs_1, dfs_2])
-    assert list(merged_dfs.keys()) == pages
-    for page in pages:
-        pd.testing.assert_frame_equal(
-            merged_dfs[page],
-            pd.DataFrame({"df": [1, 2], "page": [page, page]}, index=[0, 1]),
-        )
